requirements.txt
# Montgomery RISE v4.0.0 — Python dependencies
# Install: pip install -r requirements.txt

# ── API framework ─────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── HTTP client — ArcGIS, Gemini 2.5 Pro, 311, Nominatim ─────────
requests>=2.31.0

# ── RAG: vector store ─────────────────────────────────────────────
chromadb>=0.5.0

# ── RAG: local embeddings (all-MiniLM-L6-v2, ~80 MB, no API key) ─
sentence-transformers>=3.0.0

# ── RAG: LLM via Groq free tier (llama-3.3-70b-versatile) ────────
# Get your free key at: https://console.groq.com
groq>=0.9.0

# ── Optional: rich terminal output for CLI ingestion script ───────
rich>=13.0.0
